new-djangoproject/djangoproject/requirements.txt
djangorestframework
tensorflow
Pillow